UIMA-143 and 144 - Refactor CAS Interface into part which can be shared in common with the JCas Interface.  Added new shared common interface class (CommonCas), having sub-interfaces CAS and JCas.  Found several missing (in interface) and wrong impls of things in the process.
Missing: getView(String local_name_of_sofa) and getView(SofaFS)
JCas impl of getView was always creating new JCas rather
than reusing existing one (if it exists).

Added additional methods to JCas with impls that just forward
to CASImpl, for things that should be common.

Found a lot of tests in the JCas to see if the JCas was 
associated with a "TCAS" style CAS.  These were removed -
because TCAS stuff has been moved to CAS.


git-svn-id: https://svn.apache.org/repos/asf/incubator/uima/uimaj/trunk@490358 13f79535-47bb-0310-9956-ffa450edef68