UIMA-205 Fix case of adding flow constraints to user-defined flows having previously no flow constraint.
Also - clarify messages about key names and import names
in user-defined flows, by showing a Warning string in those
cases where there is no value for these fields.  It is allowed
to have no key name, (but then you cannot reference that
component for specifying an overriding parameter, for instance).
It is also allowed to not have an import, but in that case
you have to specify the flow component descriptor "in-line".

git-svn-id: https://svn.apache.org/repos/asf/incubator/uima/uimaj/trunk@497442 13f79535-47bb-0310-9956-ffa450edef68