JIRA ticket UIMA-280 (https://issues.apache.org/jira/browse/UIMA-280)
add uimaj-internal-tools component. This component will host UIMA internal programs that will not be shipped with UIMA. E.g. tools that helps during the build.

The JiraIssueDetector can detect the Jira issues that were fixed since the last level/tag that was build. So it is possible to generate a list of Jira issues that will be
fixed with the current level/tag.

git-svn-id: https://svn.apache.org/repos/asf/incubator/uima/uimaj/trunk@504091 13f79535-47bb-0310-9956-ffa450edef68