UIMA-284.  3 fixes.  1) added volatile to fields in this class not otherwise under synch locks for read/writing, where the fields are set/read on different threads.  Without this, different threads can see different values for these because the JVM may keep the value of these in "registers" local to the thread.
2) Moved tests leading to wait/notify actions into the same
synch lock used for the wait/notify - to avoid a race condition.

3) In the method runSingleThreaded was a call to 
ifPaused() with the comment // blocks if CPM is paused.  But
ifPaused() just tests a variable, doesn't wait for CPM to
resume.  There is a similarly named method, ifCpmPaused(), that
is a method of ProcessingUnit. I copied this method, changing
it slightly to get it to work in CPMEngine class, and renamed it
to "waitForCpmToResumeIfPaused" which is closer to what
it does, and changed the ifPaused() call to this new method.
At some point the other method in ProcessingUnit might be deleted
and this one used in all cases (a refactoring saved for later).

git-svn-id: https://svn.apache.org/repos/asf/incubator/uima/uimaj/trunk@504902 13f79535-47bb-0310-9956-ffa450edef68