UIMA-411
change the PEAR verification implementation to use the UIMA extension classpath. Now it
is never necessary to create a new process to do the PEAR verification, so no classpath and
datapath must be constructed.

To do the modification it was necessary to change some of the signitures of public methods in
InstallationController, InstallationTester and LocalInstallationAgent. But I think these methods
were not called from an application. I think most of the users use the common public method on the
InstallationController that are described in the documentation and these methods are unchanged.

JIRA ticket https://issues.apache.org/jira/browse/UIMA-411

git-svn-id: https://svn.apache.org/repos/asf/incubator/uima/uimaj/trunk/uimaj-core@540648 13f79535-47bb-0310-9956-ffa450edef68