[UIMA-386] switching classloaders.  several fixes based on new test case: FlowContainer:  Change the arg to switchClassLoader to use resource class loader FlowControllerContainer: same as above PrimitiveAnalysisEngine: same as above
CASImpl:
Set previousJCasClassLoader to the same initial default value as the jcasClassLoader on initialization
put this value in the shared-by-all-views section
Remove call to initFSClassRegistry() when creating a view - this registry is now shared with all type systems
Make call to setJCasClassLoader also set the previous value.  This call is only used when creating the CAS, not used internally.

FSClassRegistry: comment change on how this object is shared by type system and class loader

JCasImpl:  New 2 level weak hash map to map from ts and cl to loaded JCasTypes.
LoadedJCasType: change to use type name - the Type object can be different for different CASes.
When creating new JCas, skip loadking the JCasClasses part if they're already loaded for this type system and class loader.
Add code to expand the typeArray if needed (needed when new class loader used)
If JCasTypes already loaded, don't update the FSClassRegistry - not needed., when making new instances for new views.  And no need to copydownSuperGenerators, either.
When making a new instance of _Type, use current CAS to get Type object.

Activate the new test in PearRuntimeTest

git-svn-id: https://svn.apache.org/repos/asf/incubator/uima/uimaj/trunk/uimaj-core@544569 13f79535-47bb-0310-9956-ffa450edef68