[UIMA-2211] add common shared utility code that does things in preparation for leaving the framework and entering user code (annotator calls, flow controller calls), so this can be more easily maintained.  Insure that calls to switchClassLoader are preceded by calls that instantiate a JCas (if the JCas interface is being used), to get the JCas class loaders switched by the switchClassLoaders methods in CASImpl.
Change the release code to first unlock a CAS and then do the "reset".  Update 1 affected test case.

In framework code which prepares to enter user-code, which have "finally" clauses to unlock the CAS etc., skip those if an exception happens early enough in the try block that the CAS was never locked.

git-svn-id: https://svn.apache.org/repos/asf/uima/uimaj/trunk/uimaj-core@1157122 13f79535-47bb-0310-9956-ffa450edef68