UIMA-6199 2 part fix for edge case of non-indexed doc_annot
The first part: change the deserialization code to "not" index a newly
created doc_annot, but leave a previously available one (perhaps due to
using delta deserialization).

If the newly created doc_annot was previously indexed, a later part of
the deserialization code will add it to the indexes.

In the meanwhile, if a sofa update using sofa-string data (e.g. for
delta deserialization) this causes the doc_annot to be updated with
"length" of the sofa string.  Allow this code to find the newly created
doc_annot even if it wasn't indexed, by keeping a pointer to it in the
CASImpl (reset to null using a try finally block.